refactor: move tokenization methods inside a `Tokenizer` struct type
