+++ packages/fuels-core/src/json_abi.rs
@@ -1,16 +1,16 @@
+use crate::tokenizer::Tokenizer;
 use crate::utils::first_four_bytes_of_sha256_hash;
 use crate::Token;
 use crate::{abi_decoder::ABIDecoder, abi_encoder::ABIEncoder};
 use fuels_types::function_selector::build_fn_selector;
-use fuels_types::utils::has_array_format;
 use fuels_types::{errors::Error, param_types::ParamType, JsonABI, Property};
-use hex::FromHex;
 use itertools::Itertools;
 use serde_json;
 use std::str;
 
 pub struct ABIParser {
     fn_selector: Option<Vec<u8>>,
+    tokenizer: Tokenizer,
 }
 
 impl Default for ABIParser {
@@ -21,7 +21,10 @@ impl Default for ABIParser {
 
 impl ABIParser {
     pub fn new() -> Self {
-        ABIParser { fn_selector: None }
+        ABIParser {
+            fn_selector: None,
+            tokenizer: Tokenizer {},
+        }
     }
 
     /// Higher-level layer of the ABI encoding module.
@@ -178,349 +181,9 @@ impl ABIParser {
     pub fn parse_tokens<'a>(&self, params: &'a [(ParamType, &str)]) -> Result<Vec<Token>, Error> {
         params
             .iter()
-            .map(|&(ref param, value)| self.tokenize(param, value.to_string()))
+            .map(|&(ref param, value)| self.tokenizer.tokenize(param, value.to_string()))
             .collect::<Result<_, _>>()
             .map_err(From::from)
-    }
-
-    /// Takes a ParamType and a value string and joins them as a single
-    /// Token that holds the value within it. This Token is used
-    /// in the encoding process.
-    pub fn tokenize(&self, param: &ParamType, value: String) -> Result<Token, Error> {
-        let trimmed_value = value.trim();
-        match &*param {
-            ParamType::Unit => Ok(Token::Unit),
-            ParamType::U8 => Ok(Token::U8(trimmed_value.parse::<u8>()?)),
-            ParamType::U16 => Ok(Token::U16(trimmed_value.parse::<u16>()?)),
-            ParamType::U32 => Ok(Token::U32(trimmed_value.parse::<u32>()?)),
-            ParamType::U64 => Ok(Token::U64(trimmed_value.parse::<u64>()?)),
-            ParamType::Bool => Ok(Token::Bool(trimmed_value.parse::<bool>()?)),
-            ParamType::Byte => Ok(Token::Byte(trimmed_value.parse::<u8>()?)),
-            ParamType::B256 => {
-                const B256_HEX_ENC_LENGTH: usize = 64;
-                if trimmed_value.len() != B256_HEX_ENC_LENGTH {
-                    return Err(Error::InvalidData(format!(
-                        "the hex encoding of the b256 must have {} characters",
-                        B256_HEX_ENC_LENGTH
-                    )));
-                }
-                let v = Vec::from_hex(trimmed_value)?;
-                let s: [u8; 32] = v.as_slice().try_into().unwrap();
-                Ok(Token::B256(s))
-            }
-            ParamType::Array(t, _) => Ok(self.tokenize_array(trimmed_value, &*t)?),
-            ParamType::String(_) => Ok(Token::String(trimmed_value.to_string())),
-            ParamType::Struct(struct_params) => {
-                Ok(self.tokenize_struct(trimmed_value, struct_params)?)
-            }
-            ParamType::Enum(variants) => {
-                let discriminant = self.get_enum_discriminant_from_string(trimmed_value);
-                let value = self.get_enum_value_from_string(trimmed_value);
-
-                let token = self.tokenize(&variants.param_types()[discriminant], value)?;
-
-                Ok(Token::Enum(Box::new((
-                    discriminant as u8,
-                    token,
-                    variants.clone(),
-                ))))
-            }
-            ParamType::Tuple(tuple_params) => Ok(self.tokenize_tuple(trimmed_value, tuple_params)?),
-        }
-    }
-
-    /// Creates a `Token::Struct` from an array of parameter types and a string of values.
-    /// I.e. it takes a string containing values "value_1, value_2, value_3" and an array
-    /// of `ParamType` containing the type of each value, in order:
-    /// [ParamType::<Type of value_1>, ParamType::<Type of value_2>, ParamType::<Type of value_3>]
-    /// And attempts to return a `Token::Struct()` containing the inner types.
-    /// It works for nested/recursive structs.
-    pub fn tokenize_struct(&self, value: &str, params: &[ParamType]) -> Result<Token, Error> {
-        if !value.starts_with('(') || !value.ends_with(')') {
-            return Err(Error::InvalidData(
-                "struct value string must start and end with round brackets".into(),
-            ));
-        }
-
-        if value.chars().count() == 2 {
-            return Ok(Token::Struct(vec![]));
-        }
-
-        //To parse the value string we use a two pointer/index approach.
-        //The items are comma separated and if an item is tokenized the last_item
-        //index is moved to the current position.
-        //The variable nested is incremented and decremented if a bracket is encountered,
-        //and appropriate errors are returned if the nested count is not 0.
-        //If the struct has an array inside its values the current position will be incremented
-        //until both the opening and closing bracket are inside the new item.
-        //Characters inside quotes are ignored and they are tokenized as one item.
-        //An error is return if there is an odd number of quotes.
-        let mut result = vec![];
-        let mut nested = 0isize;
-        let mut ignore = false;
-        let mut last_item = 1;
-        let mut params_iter = params.iter();
-
-        for (pos, ch) in value.chars().enumerate() {
-            match ch {
-                '(' if !ignore => {
-                    nested += 1;
-                }
-                ')' if !ignore => {
-                    nested -= 1;
-
-                    match nested.cmp(&0) {
-                        std::cmp::Ordering::Less => {
-                            return Err(Error::InvalidData(
-                                "struct value string has excess closing brackets".into(),
-                            ));
-                        }
-                        std::cmp::Ordering::Equal => {
-                            let sub = &value[last_item..pos];
-
-                            let token = self.tokenize(
-                                params_iter.next().ok_or_else(|| {
-                                    Error::InvalidData(
-                                        "struct value contains more elements than the parameter types provided".into(),
-                                    )
-                                })?,
-                                sub.to_string(),
-                            )?;
-                            result.push(token);
-                            last_item = pos + 1;
-                        }
-                        _ => {}
-                    }
-                }
-                '"' => {
-                    ignore = !ignore;
-                }
-                ',' if nested == 1 && !ignore => {
-                    let sub = &value[last_item..pos];
-                    // If we've encountered an array within a struct property
-                    // keep iterating until we see the end of it "]".
-                    if sub.contains('[') && !sub.contains(']') {
-                        continue;
-                    }
-
-                    let token = self.tokenize(
-                        params_iter.next().ok_or_else(|| {
-                            Error::InvalidData(
-                                "struct value contains more elements than the parameter types provided".into(),
-                            )
-                        })?,
-                        sub.to_string(),
-                    )?;
-                    result.push(token);
-                    last_item = pos + 1;
-                }
-                _ => (),
-            }
-        }
-
-        if ignore {
-            return Err(Error::InvalidData(
-                "struct value string has excess quotes".into(),
-            ));
-        }
-
-        if nested > 0 {
-            return Err(Error::InvalidData(
-                "struct value string has excess opening brackets".into(),
-            ));
-        }
-
-        Ok(Token::Struct(result))
-    }
-
-    /// Creates a `Token::Array` from one parameter type and a string of values. I.e. it takes a
-    /// string containing values "value_1, value_2, value_3" and a `ParamType` sepecifying the type.
-    /// It works for nested/recursive arrays.
-    pub fn tokenize_array<'a>(&self, value: &'a str, param: &ParamType) -> Result<Token, Error> {
-        if !value.starts_with('[') || !value.ends_with(']') {
-            return Err(Error::InvalidData(
-                "array value string must start and end with square brackets".into(),
-            ));
-        }
-
-        if value.chars().count() == 2 {
-            return Ok(Token::Array(vec![]));
-        }
-
-        //for more details about this algorithm, refer to the tokenize_struct method
-        let mut result = vec![];
-        let mut nested = 0isize;
-        let mut ignore = false;
-        let mut last_item = 1;
-        for (i, ch) in value.chars().enumerate() {
-            match ch {
-                '[' if !ignore => {
-                    nested += 1;
-                }
-                ']' if !ignore => {
-                    nested -= 1;
-
-                    match nested.cmp(&0) {
-                        std::cmp::Ordering::Less => {
-                            return Err(Error::InvalidData(
-                                "array value string has excess closing brackets".into(),
-                            ));
-                        }
-                        std::cmp::Ordering::Equal => {
-                            // Last element of this nest level; proceed to tokenize.
-                            let sub = &value[last_item..i];
-                            match has_array_format(sub) {
-                                true => {
-                                    let arr_param = ParamType::Array(
-                                        Box::new(param.to_owned()),
-                                        self.get_array_length_from_string(sub),
-                                    );
-
-                                    result.push(self.tokenize(&arr_param, sub.to_string())?);
-                                }
-                                false => {
-                                    result.push(self.tokenize(param, sub.to_string())?);
-                                }
-                            }
-
-                            last_item = i + 1;
-                        }
-                        _ => {}
-                    }
-                }
-                '"' => {
-                    ignore = !ignore;
-                }
-                ',' if nested == 1 && !ignore => {
-                    let sub = &value[last_item..i];
-                    match has_array_format(sub) {
-                        true => {
-                            let arr_param = ParamType::Array(
-                                Box::new(param.to_owned()),
-                                self.get_array_length_from_string(sub),
-                            );
-
-                            result.push(self.tokenize(&arr_param, sub.to_string())?);
-                        }
-                        false => {
-                            result.push(self.tokenize(param, sub.to_string())?);
-                        }
-                    }
-                    last_item = i + 1;
-                }
-                _ => (),
-            }
-        }
-
-        if ignore {
-            return Err(Error::InvalidData(
-                "array value string has excess quotes".into(),
-            ));
-        }
-
-        if nested > 0 {
-            return Err(Error::InvalidData(
-                "array value string has excess opening brackets".into(),
-            ));
-        }
-
-        Ok(Token::Array(result))
-    }
-
-    /// Creates `Token::Tuple` from an array of parameter types and a string of values.
-    /// I.e. it takes a string containing values "value_1, value_2, value_3" and an array
-    /// of `ParamType` containing the type of each value, in order:
-    /// [ParamType::<Type of value_1>, ParamType::<Type of value_2>, ParamType::<Type of value_3>]
-    /// And attempts to return a `Token::Tuple()` containing the inner types.
-    /// It works for nested/recursive tuples.
-    pub fn tokenize_tuple(&self, value: &str, params: &[ParamType]) -> Result<Token, Error> {
-        if !value.starts_with('(') || !value.ends_with(')') {
-            return Err(Error::InvalidData(
-                "tuple value string must start and end with round brackets".into(),
-            ));
-        }
-
-        if value.chars().count() == 2 {
-            return Ok(Token::Tuple(vec![]));
-        }
-
-        //for more details about this algorithm, refer to the tokenize_struct method
-        let mut result = vec![];
-        let mut nested = 0isize;
-        let mut ignore = false;
-        let mut last_item = 1;
-        let mut params_iter = params.iter();
-
-        for (pos, ch) in value.chars().enumerate() {
-            match ch {
-                '(' if !ignore => {
-                    nested += 1;
-                }
-                ')' if !ignore => {
-                    nested -= 1;
-
-                    match nested.cmp(&0) {
-                        std::cmp::Ordering::Less => {
-                            return Err(Error::InvalidData(
-                                "tuple value string has excess closing brackets".into(),
-                            ));
-                        }
-                        std::cmp::Ordering::Equal => {
-                            let sub = &value[last_item..pos];
-
-                            let token = self.tokenize(
-                                params_iter.next().ok_or_else(|| {
-                                    Error::InvalidData(
-                                        "tuple value contains more elements than the parameter types provided".into(),
-                                    )
-                                })?,
-                                sub.to_string(),
-                            )?;
-                            result.push(token);
-                            last_item = pos + 1;
-                        }
-                        _ => {}
-                    }
-                }
-                '"' => {
-                    ignore = !ignore;
-                }
-                ',' if nested == 1 && !ignore => {
-                    let sub = &value[last_item..pos];
-                    // If we've encountered an array within a tuple property
-                    // keep iterating until we see the end of it "]".
-                    if sub.contains('[') && !sub.contains(']') {
-                        continue;
-                    }
-
-                    let token = self.tokenize(
-                        params_iter.next().ok_or_else(|| {
-                            Error::InvalidData(
-                                "tuple value contains more elements than the parameter types provided".into(),
-                            )
-                        })?,
-                        sub.to_string(),
-                    )?;
-                    result.push(token);
-                    last_item = pos + 1;
-                }
-                _ => (),
-            }
-        }
-
-        if ignore {
-            return Err(Error::InvalidData(
-                "tuple value string has excess quotes".into(),
-            ));
-        }
-
-        if nested > 0 {
-            return Err(Error::InvalidData(
-                "tuple value string has excess opening brackets".into(),
-            ));
-        }
-
-        Ok(Token::Tuple(result))
     }
 
     /// Higher-level layer of the ABI decoding module.
@@ -561,29 +224,6 @@ impl ABIParser {
     /// without having to reference to a JSON specification of the ABI.
     pub fn decode_params(&self, params: &[ParamType], data: &[u8]) -> Result<Vec<Token>, Error> {
         Ok(ABIDecoder::decode(params, data)?)
-    }
-
-    fn get_enum_discriminant_from_string(&self, ele: &str) -> usize {
-        let mut chars = ele.chars();
-        chars.next(); // Remove "("
-        chars.next_back(); // Remove ")"
-        let v: Vec<_> = chars.as_str().split(',').collect();
-        v[0].parse().unwrap()
-    }
-
-    fn get_enum_value_from_string(&self, ele: &str) -> String {
-        let mut chars = ele.chars();
-        chars.next(); // Remove "("
-        chars.next_back(); // Remove ")"
-        let v: Vec<_> = chars.as_str().split(',').collect();
-        v[1].to_string()
-    }
-
-    fn get_array_length_from_string(&self, ele: &str) -> usize {
-        let mut chars = ele.chars();
-        chars.next();
-        chars.next_back();
-        chars.as_str().split(',').count()
     }
 }
 
@@ -787,11 +427,11 @@ mod tests {
 
     #[test]
     fn tokenize_array() -> Result<(), Error> {
-        let abi = ABIParser::new();
+        let tokenizer = Tokenizer::new();
 
         let value = "[[1,2],[3],4]";
         let param = ParamType::U16;
-        let tokens = abi.tokenize_array(value, &param)?;
+        let tokens = tokenizer.tokenize_array(value, &param)?;
 
         let expected_tokens = Token::Array(vec![
             Token::Array(vec![Token::U16(1), Token::U16(2)]), // First element, a sub-array with 2 elements
@@ -803,7 +443,7 @@ mod tests {
 
         let value = "[1,[2],[3],[4,5]]";
         let param = ParamType::U16;
-        let tokens = abi.tokenize_array(value, &param)?;
+        let tokens = tokenizer.tokenize_array(value, &param)?;
 
         let expected_tokens = Token::Array(vec![
             Token::U16(1),
@@ -816,7 +456,7 @@ mod tests {
 
         let value = "[1,2,3,4,5]";
         let param = ParamType::U16;
-        let tokens = abi.tokenize_array(value, &param)?;
+        let tokens = tokenizer.tokenize_array(value, &param)?;
 
         let expected_tokens = Token::Array(vec![
             Token::U16(1),
@@ -830,7 +470,7 @@ mod tests {
 
         let value = "[[1,2,3,[4,5]]]";
         let param = ParamType::U16;
-        let tokens = abi.tokenize_array(value, &param)?;
+        let tokens = tokenizer.tokenize_array(value, &param)?;
 
         let expected_tokens = Token::Array(vec![Token::Array(vec![
             Token::U16(1),
@@ -1560,490 +1200,6 @@ mod tests {
 
         let expected_encode = "00000000e33a11ce0000000000000001000000000000002a";
         assert_eq!(encoded, expected_encode);
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_uint_types_expected_error() {
-        let abi = ABIParser::new();
-
-        // We test only on U8 as it is the same error on all other unsigned int types
-        let error_message = abi
-            .tokenize(&ParamType::U8, "2,".to_string())
-            .unwrap_err()
-            .to_string();
-
-        assert_eq!(
-            "Parse integer error: invalid digit found in string",
-            error_message
-        );
-    }
-
-    #[test]
-    fn tokenize_bool_expected_error() {
-        let abi = ABIParser::new();
-
-        let error_message = abi
-            .tokenize(&ParamType::Bool, "True".to_string())
-            .unwrap_err()
-            .to_string();
-
-        assert_eq!(
-            "Parse boolean error: provided string was not `true` or `false`",
-            error_message
-        );
-    }
-
-    #[test]
-    fn tokenize_b256_invalid_length_expected_error() {
-        let abi = ABIParser::new();
-
-        let value = "d57a9c46dfcc7f18207013e65b44e4cb4e2c2298f4ac457ba8f82743f31e90b".to_string();
-        let error_message = abi
-            .tokenize(&ParamType::B256, value)
-            .unwrap_err()
-            .to_string();
-
-        assert_eq!(
-            "Invalid data: the hex encoding of the b256 must have 64 characters",
-            error_message
-        );
-    }
-
-    #[test]
-    fn tokenize_b256_invalid_character_expected_error() {
-        let abi = ABIParser::new();
-
-        let value = "Hd57a9c46dfcc7f18207013e65b44e4cb4e2c2298f4ac457ba8f82743f31e90b".to_string();
-        let error_message = abi
-            .tokenize(&ParamType::B256, value)
-            .unwrap_err()
-            .to_string();
-
-        assert!(error_message.contains("Parse hex error: Invalid character"));
-    }
-
-    #[test]
-    fn tokenize_tuple_invalid_start_end_bracket_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "(u64, [u64; 3])".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Tuple(tuple_params) = ParamType::parse_tuple_param(&params)? {
-            let error_message = abi
-                .tokenize_tuple("0, [0,0,0])", &tuple_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: tuple value string must start and end with round brackets",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_tuple_excess_opening_bracket_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "(u64, [u64; 3])".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Tuple(tuple_params) = ParamType::parse_tuple_param(&params)? {
-            let error_message = abi
-                .tokenize_tuple("((0, [0,0,0])", &tuple_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: tuple value string has excess opening brackets",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_tuple_excess_closing_bracket_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "(u64, [u64; 3])".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Tuple(tuple_params) = ParamType::parse_tuple_param(&params)? {
-            let error_message = abi
-                .tokenize_tuple("(0, [0,0,0]))", &tuple_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: tuple value string has excess closing brackets",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_tuple_excess_quotes_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "(u64, [u64; 3])".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Tuple(tuple_params) = ParamType::parse_tuple_param(&params)? {
-            let error_message = abi
-                .tokenize_tuple("(0, \"[0,0,0])", &tuple_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: tuple value string has excess quotes",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_tuple_excess_value_elements_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "(u64, [u64; 3])".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "__tuple_element".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Tuple(tuple_params) = ParamType::parse_tuple_param(&params)? {
-            let error_message = abi
-                .tokenize_tuple("(0, [0,0,0], 0, 0)", &tuple_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: tuple value contains more elements than the parameter types provided",
-                error_message
-            );
-
-            let error_message = abi
-                .tokenize_tuple("(0, [0,0,0], 0)", &tuple_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: tuple value contains more elements than the parameter types provided",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_array_invalid_start_end_bracket_expected_error() {
-        let param = ParamType::U16;
-        let abi = ABIParser::new();
-
-        let error_message = abi
-            .tokenize_array("1,2],[3],4]", &param)
-            .unwrap_err()
-            .to_string();
-
-        assert_eq!(
-            "Invalid data: array value string must start and end with square brackets",
-            error_message
-        );
-    }
-
-    #[test]
-    fn tokenize_array_excess_opening_bracket_expected_error() {
-        let param = ParamType::U16;
-        let abi = ABIParser::new();
-
-        let error_message = abi
-            .tokenize_array("[[[1,2],[3],4]", &param)
-            .unwrap_err()
-            .to_string();
-
-        assert_eq!(
-            "Invalid data: array value string has excess opening brackets",
-            error_message
-        );
-    }
-
-    #[test]
-    fn tokenize_array_excess_closing_bracket_expected_error() {
-        let param = ParamType::U16;
-        let abi = ABIParser::new();
-
-        let error_message = abi
-            .tokenize_array("[[1,2],[3],4]]", &param)
-            .unwrap_err()
-            .to_string();
-
-        assert_eq!(
-            "Invalid data: array value string has excess closing brackets",
-            error_message
-        );
-    }
-
-    #[test]
-    fn tokenize_array_excess_quotes_expected_error() {
-        let param = ParamType::U16;
-        let abi = ABIParser::new();
-
-        let error_message = abi
-            .tokenize_array("[[1,\"2],[3],4]]", &param)
-            .unwrap_err()
-            .to_string();
-
-        assert_eq!(
-            "Invalid data: array value string has excess quotes",
-            error_message
-        );
-    }
-
-    #[test]
-    fn tokenize_struct_invalid_start_end_bracket_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "struct MyStruct".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "num".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "arr".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Struct(struct_params) = ParamType::parse_custom_type_param(&params)? {
-            let error_message = abi
-                .tokenize_struct("0, [0,0,0])", &struct_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: struct value string must start and end with round brackets",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_struct_excess_opening_bracket_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "struct MyStruct".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "num".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "arr".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Struct(struct_params) = ParamType::parse_custom_type_param(&params)? {
-            let error_message = abi
-                .tokenize_struct("((0, [0,0,0])", &struct_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: struct value string has excess opening brackets",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_struct_excess_closing_bracket_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "struct MyStruct".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "num".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "arr".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Struct(struct_params) = ParamType::parse_custom_type_param(&params)? {
-            let error_message = abi
-                .tokenize_struct("(0, [0,0,0]))", &struct_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: struct value string has excess closing brackets",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_struct_excess_quotes_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "struct MyStruct".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "num".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "arr".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Struct(struct_params) = ParamType::parse_custom_type_param(&params)? {
-            let error_message = abi
-                .tokenize_struct("(0, \"[0,0,0])", &struct_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: struct value string has excess quotes",
-                error_message
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn tokenize_struct_excess_value_elements_expected_error() -> Result<(), Error> {
-        let abi = ABIParser::new();
-        let params = Property {
-            name: "input".to_string(),
-            type_field: "struct MyStruct".to_string(),
-            components: Some(vec![
-                Property {
-                    name: "num".to_string(),
-                    type_field: "u64".to_string(),
-                    components: None,
-                },
-                Property {
-                    name: "arr".to_string(),
-                    type_field: "[u64; 3]".to_string(),
-                    components: None,
-                },
-            ]),
-        };
-
-        if let ParamType::Struct(struct_params) = ParamType::parse_custom_type_param(&params)? {
-            let error_message = abi
-                .tokenize_struct("(0, [0,0,0], 0, 0)", &struct_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: struct value contains more elements than the parameter types provided",
-                error_message
-            );
-
-            let error_message = abi
-                .tokenize_struct("(0, [0,0,0], 0)", &struct_params)
-                .unwrap_err()
-                .to_string();
-
-            assert_eq!(
-                "Invalid data: struct value contains more elements than the parameter types provided",
-                error_message
-            );
-        }
         Ok(())
     }
 }
